printer_daemon/pyproject.toml
[tool.poetry]
name = "printer-daemon"
version = "0.1.0"
description = "Daemon to receive MQTT messages and print them to a serial printer."
authors = ["Cline <cline@example.com>"]
readme = "README.md"
packages = [{include = "printer_daemon", from = "src"}]

[tool.poetry.dependencies]
python = "^3.12"
paho-mqtt = "^1.6.1"
pyserial = "^3.5"

[tool.poetry.group.dev.dependencies]
pytest = "^7.0"
pytest-cov = "^4.0"
black = "^23.0"
flake8 = "^6.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
